Initial Commit.
This is the entire Button Shooter application. The applications starts
first with a menu with two options first: Play game. Second: quit. The
game has a button when pressed changes text just above the button to
"PEW!" and every now and then the user will score a hit. When that
happens the user is informed by a randomly colored "Its a hit!".